keep external file for the handler
